feat(internal/adapters): Relocate dataset handler
* Relocated the dataset export worker and HTTP handler into the adapters layer
* Relocated the dataset package to `datasets`
* Re-pointed adapter tests to the new import path and kept the in-memory helpers available for consumers
* Dropped the obsolete internal/dataset tree
* Confirmed all references use the new adapter location
* Updated documentation to reference the new dataset adapter path
