+++ internal/adapters/datasets/exporter.go
@@ -1,4 +1,4 @@
-package dataset
+package datasets
 
 import (
 	"bytes"
