+++ internal/adapters/datasets/exporter_materialize_test.go
@@ -1,4 +1,4 @@
-package dataset
+package datasets
 
 import (
 	"colonycore/internal/core"
