+++ internal/adapters/datasets/handler.go
@@ -1,4 +1,4 @@
-package dataset
+package datasets
 
 import (
 	"encoding/csv"
