+++ internal/adapters/datasets/handler_internal_test.go
@@ -1,4 +1,4 @@
-package dataset
+package datasets
 
 import (
 	"bytes"
