+++ internal/adapters/datasets/objectstore_test.go
@@ -1,4 +1,4 @@
-package dataset
+package datasets
 
 import (
 	"context"
